+++ Source/LambdaSharp.DynamoDB.Native/Operations/IDynamoTableTransactWriteItems.cs
@@ -111,12 +111,12 @@ namespace LambdaSharp.DynamoDB.Native.Operations {
         /// <summary>
         /// Add condition that item exists for PutItem operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsPutItem<TRecord> WithConditionExists() => WithCondition(record => DynamoCondition.Exists(record));
+        IDynamoTableTransactWriteItemsPutItem<TRecord> WithConditionItemExists() => WithCondition(record => DynamoCondition.Exists(record));
 
         /// <summary>
         /// Add condition that item does not exist for PutItem operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsPutItem<TRecord> WithConditionDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
+        IDynamoTableTransactWriteItemsPutItem<TRecord> WithConditionItemDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
 
         /// <summary>
         /// Set the value of a DynamoDB item attribute. Used for storing attributes used by local/global secondary indices.
@@ -148,12 +148,12 @@ namespace LambdaSharp.DynamoDB.Native.Operations {
         /// <summary>
         /// Add condition that item exists for UpdateItem operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsUpdateItem<TRecord> WithConditionExists() => WithCondition(record => DynamoCondition.Exists(record));
+        IDynamoTableTransactWriteItemsUpdateItem<TRecord> WithConditionItemExists() => WithCondition(record => DynamoCondition.Exists(record));
 
         /// <summary>
         /// Add condition that item does not exist for UpdateItem operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsUpdateItem<TRecord> WithConditionDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
+        IDynamoTableTransactWriteItemsUpdateItem<TRecord> WithConditionItemDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
 
         /// <summary>
         /// End specification of the UpdateItem operation for TransactWriteItems.
@@ -392,12 +392,12 @@ namespace LambdaSharp.DynamoDB.Native.Operations {
         /// <summary>
         /// Add condition that item exists for DeleteItem operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsDeleteItem<TRecord> WithConditionExists() => WithCondition(record => DynamoCondition.Exists(record));
+        IDynamoTableTransactWriteItemsDeleteItem<TRecord> WithConditionItemExists() => WithCondition(record => DynamoCondition.Exists(record));
 
         /// <summary>
         /// Add condition that item does not exist for DeleteItem operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsDeleteItem<TRecord> WithConditionDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
+        IDynamoTableTransactWriteItemsDeleteItem<TRecord> WithConditionItemDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
 
         /// <summary>
         /// End specification of the DeleteItem operation for TransactWriteItems.
@@ -422,12 +422,12 @@ namespace LambdaSharp.DynamoDB.Native.Operations {
         /// <summary>
         /// Add condition that item exists for ConditionCheck operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsConditionCheck<TRecord> WithConditionExists() => WithCondition(record => DynamoCondition.Exists(record));
+        IDynamoTableTransactWriteItemsConditionCheck<TRecord> WithConditionItemExists() => WithCondition(record => DynamoCondition.Exists(record));
 
         /// <summary>
         /// Add condition that item does not exist for ConditionCheck operation.
         /// </summary>
-        IDynamoTableTransactWriteItemsConditionCheck<TRecord> WithConditionDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
+        IDynamoTableTransactWriteItemsConditionCheck<TRecord> WithConditionItemDoesNotExist() => WithCondition(record => DynamoCondition.DoesNotExist(record));
 
         /// <summary>
         /// End specification of the ConditionCheck operation for TransactWriteItems.