+++ ReleaseNotes.md
@@ -7,5 +7,5 @@
     * Migrated assembly from [LambdaSharpTool](https://github.com/LambdaSharp/LambdaSharpTool) repository to [LambdaSharp.DynamoDB](https://github.com/LambdaSharp/LambdaSharp.DynamoDB/) repository.
 
 * SDK
-    * Added `WithConditionExists()` as shortcut for `WithCondition(record => DynamoCondition.Exists(record))`.
-    * Added `WithConditionDoesNotExist()` as shortcut for `WithCondition(record => DynamoCondition.DoesNotExist(record))`.
+    * Added `WithConditionItemExists()` as shortcut for `WithCondition(record => DynamoCondition.Exists(record))`.
+    * Added `WithConditionItemDoesNotExist()` as shortcut for `WithCondition(record => DynamoCondition.DoesNotExist(record))`.